Dev UI - qute - improve the "render preview" page
- use codemirror editor for test data
- prefill the test data for type-safe templates
- show the source of the template
