+++ extensions/qute/deployment/src/main/java/io/quarkus/qute/deployment/QuteProcessor.java
@@ -1,6 +1,5 @@
 package io.quarkus.qute.deployment;
 
-import static io.netty.handler.codec.http.HttpHeaderNames.CONTENT_TYPE;
 import static io.quarkus.deployment.annotations.ExecutionTime.STATIC_INIT;
 import static java.util.stream.Collectors.toMap;
 
@@ -9,7 +8,6 @@ import java.io.IOException;
 import java.io.Reader;
 import java.io.StringReader;
 import java.lang.reflect.Modifier;
-import java.net.URLConnection;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -27,7 +25,6 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.CompletionStage;
-import java.util.function.BiFunction;
 import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.regex.Pattern;
@@ -78,8 +75,6 @@ import io.quarkus.deployment.builditem.ServiceStartBuildItem;
 import io.quarkus.deployment.builditem.nativeimage.NativeImageResourceBuildItem;
 import io.quarkus.deployment.builditem.nativeimage.ReflectiveClassBuildItem;
 import io.quarkus.deployment.util.JandexUtil;
-import io.quarkus.dev.console.DevConsoleManager;
-import io.quarkus.devconsole.spi.DevConsoleRouteBuildItem;
 import io.quarkus.gizmo.ClassOutput;
 import io.quarkus.panache.common.deployment.PanacheEntityClassesBuildItem;
 import io.quarkus.qute.CheckedTemplate;
@@ -116,20 +111,12 @@ import io.quarkus.qute.runtime.QuteConfig;
 import io.quarkus.qute.runtime.QuteRecorder;
 import io.quarkus.qute.runtime.QuteRecorder.QuteContext;
 import io.quarkus.qute.runtime.TemplateProducer;
-import io.quarkus.qute.runtime.devmode.QuteDevConsoleRecorder;
 import io.quarkus.qute.runtime.extensions.CollectionTemplateExtensions;
 import io.quarkus.qute.runtime.extensions.ConfigTemplateExtensions;
 import io.quarkus.qute.runtime.extensions.MapTemplateExtensions;
 import io.quarkus.qute.runtime.extensions.NumberTemplateExtensions;
 import io.quarkus.qute.runtime.extensions.StringTemplateExtensions;
 import io.quarkus.qute.runtime.extensions.TimeTemplateExtensions;
-import io.vertx.core.Handler;
-import io.vertx.core.MultiMap;
-import io.vertx.core.json.DecodeException;
-import io.vertx.core.json.Json;
-import io.vertx.core.json.JsonArray;
-import io.vertx.core.json.JsonObject;
-import io.vertx.ext.web.RoutingContext;
 
 public class QuteProcessor {
 
@@ -1236,54 +1223,6 @@ public class QuteProcessor {
     }
 
     @BuildStep
-    @Record(value = STATIC_INIT, optional = true)
-    DevConsoleRouteBuildItem invokeEndpoint(QuteDevConsoleRecorder recorder) {
-        recorder.setupRenderer();
-        return new DevConsoleRouteBuildItem("preview", "POST", new Handler<RoutingContext>() {
-            @Override
-            public void handle(RoutingContext context) {
-                context.request().setExpectMultipart(true);
-                context.request().endHandler(new Handler<Void>() {
-                    @Override
-                    public void handle(Void ignore) {
-                        MultiMap form = context.request().formAttributes();
-                        String templatePath = form.get("template-path");
-                        String testJsonData = form.get("template-data");
-                        String contentType = null;
-                        String fileName = templatePath;
-                        int slashIdx = fileName.lastIndexOf('/');
-                        if (slashIdx != -1) {
-                            fileName = fileName.substring(slashIdx, fileName.length());
-                        }
-                        int dotIdx = fileName.lastIndexOf('.');
-                        if (dotIdx != -1) {
-                            String suffix = fileName.substring(dotIdx + 1, fileName.length());
-                            if (suffix.equalsIgnoreCase("json")) {
-                                contentType = Variant.APPLICATION_JSON;
-                            } else {
-                                contentType = URLConnection.getFileNameMap().getContentTypeFor(fileName);
-                            }
-                        }
-                        try {
-                            BiFunction<String, Object, String> renderer = DevConsoleManager
-                                    .getGlobal(QuteDevConsoleRecorder.RENDER_HANDLER);
-                            Object testData = Json.decodeValue(testJsonData);
-                            testData = translate(testData); //translate it to JDK types
-                            context.response().setStatusCode(200).putHeader(CONTENT_TYPE, contentType)
-                                    .end(renderer.apply(templatePath, testData));
-                        } catch (DecodeException e) {
-                            context.response().setStatusCode(500).putHeader(CONTENT_TYPE, "text/plain; charset=UTF-8")
-                                    .end("Failed to parse JSON: " + e.getMessage());
-                        } catch (Throwable e) {
-                            context.fail(e);
-                        }
-                    }
-                });
-            }
-        });
-    }
-
-    @BuildStep
     QualifierRegistrarBuildItem turnLocationIntoQualifier() {
         return new QualifierRegistrarBuildItem(new QualifierRegistrar() {
 
@@ -1292,38 +1231,6 @@ public class QuteProcessor {
                 return Collections.singletonMap(Names.LOCATION, Collections.singleton("value"));
             }
         });
-    }
-
-    /**
-     * translates Json types to JDK types
-     *
-     * @param testData
-     * @return
-     */
-    private Object translate(Object testData) {
-        if (testData instanceof JsonArray) {
-            return translate((JsonArray) testData);
-        } else if (testData instanceof JsonObject) {
-            return translate((JsonObject) testData);
-        }
-        return testData;
-    }
-
-    private Object translate(JsonArray testData) {
-        List<Object> ret = new ArrayList<>();
-        for (Object i : testData.getList()) {
-            ret.add(translate(i));
-        }
-        return ret;
-    }
-
-    private Object translate(JsonObject testData) {
-        Map<String, Object> map = new HashMap<>();
-        Map<String, Object> data = testData.getMap();
-        for (String i : testData.fieldNames()) {
-            map.put(i, translate(data.get(i)));
-        }
-        return map;
     }
 
     private static Type resolveType(AnnotationTarget member, Match match, IndexView index,
@@ -1339,9 +1246,11 @@ public class QuteProcessor {
         // If needed attempt to resolve the type variables using the declaring type
         if (Types.containsTypeVariable(matchType)) {
             // First get the type closure of the current match type
-            Set<Type> closure = Types.getTypeClosure(match.clazz, Types.buildResolvedMap(
-                    match.getParameterizedTypeArguments(), match.getTypeParameters(),
-                    new HashMap<>(), index), index);
+            Set<Type> closure = Types.getTypeClosure(match.clazz,
+                    Types.buildResolvedMap(
+                            match.getParameterizedTypeArguments(), match.getTypeParameters(),
+                            new HashMap<>(), index),
+                    index);
 
             DotName declaringClassName = null;
             Type extensionMatchBase = null;
