+++ extensions/qute/deployment/src/test/java/io/quarkus/qute/deployment/devconsole/DevConsolePreviewTest.java
@@ -21,7 +21,7 @@ public class DevConsolePreviewTest {
 
     @Test
     public void testLoopPreview() {
-        RestAssured.with().formParam("template-path", "loop.txt").formParam("template-data", "{\"total\": [1 ,2 ,3]}")
+        RestAssured.with().formParam("template-select", "loop.txt").formParam("template-data", "{\"total\": [1 ,2 ,3]}")
                 .post("q/dev/io.quarkus.quarkus-qute/preview")
                 .then()
                 .statusCode(200)
