+++ extensions/vertx-http/deployment/pom.xml
@@ -239,6 +239,18 @@
                                     <version>${webjar.codemirror.version}</version>
                                     <type>jar</type>
                                     <overWrite>true</overWrite>
+                                    <outputDirectory>${project.build.directory}/classes/dev-static/codemirror/mode/javascript</outputDirectory>
+                                    <includes>**/mode/javascript/**.js</includes>
+                                    <fileMappers>
+                                        <org.codehaus.plexus.components.io.filemappers.FlattenFileMapper/>
+                                    </fileMappers>
+                                </artifactItem>
+                                <artifactItem>
+                                    <groupId>org.webjars</groupId>
+                                    <artifactId>codemirror</artifactId>
+                                    <version>${webjar.codemirror.version}</version>
+                                    <type>jar</type>
+                                    <overWrite>true</overWrite>
                                     <outputDirectory>${project.build.directory}/classes/dev-static/codemirror/addon/hint</outputDirectory>
                                     <includes>**/addon/hint/show-hint.js, **/addon/hint/show-hint.css</includes>
                                     <fileMappers>
